requirements.txt
requests
PyYAML
rich
prometheus_client